+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/visualize/CFGVisualizeLauncher.java
@@ -31,8 +31,7 @@ import org.plumelib.util.ArrayMap;
  *
  * <p>Usage: Directly run it as the main class to generate the DOT representation of the control
  * flow graph of a given method in a given class. See {@link
- * org.checkerframework.dataflow.cfg.playground.ConstantPropagationPlayground} for another way to
- * use it.
+ * org.checkerframework.dataflow.cfg.playground.ConstantPropagationPdf} for another way to use it.
  */
 public final class CFGVisualizeLauncher {