+++ docs/CHANGELOG.md
@@ -9,6 +9,9 @@ In `CFGVisualizeOptions`:
  * renamed `isPDF()` to `isPdfOutput()`
  * renamed `isString()` to `isStringOutput()`
 
+Dataflow Framework:
+ * In `playground` package, renamed `*Playground.java` to `*Pdf.java`
+
 **Closed issues:**